Hoist the NetworkIF enumeration loop into AbstractNetworkIF
Every platform's getNetworks(boolean) hand-rolled the same loop: enumerate
getNetworkInterfaces(includeLocal), construct the platform-specific NetworkIF
for each, and log+skip any that fail to instantiate. Only the constructed type
(and, for Mac/Aix, some pre-queried bulk data) differed.

SolarisNetworkIF already had a local NetworkIFFactory + getNetworks(boolean,
factory) helper; promote it to AbstractNetworkIF and adopt it everywhere. The
simple platforms delegate with a constructor reference; Mac and AIX capture
their pre-queried data (interface stats) in the factory lambda. Solaris's local
copy is removed in favor of the inherited one.

Along the way this unifies an inconsistency: LinuxNetworkIFFFM and
LinuxNetworkIfNF logged e.toString() while every other platform logged
e.getMessage(); the shared helper uses getMessage() for all. Now-orphaned LOG
fields are dropped.

No user-facing change. NetworkIF JNA==FFM comparison passes.
