+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/solaris/SolarisNetworkIF.java
@@ -5,14 +5,8 @@
 package oshi.hardware.common.platform.unix.solaris;
 
 import java.net.NetworkInterface;
-import java.util.ArrayList;
-import java.util.List;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.hardware.NetworkIF;
 import oshi.hardware.common.AbstractNetworkIF;
 
 /**
@@ -21,8 +15,6 @@ import oshi.hardware.common.AbstractNetworkIF;
  */
 @ThreadSafe
 public abstract class SolarisNetworkIF extends AbstractNetworkIF {
-
-    private static final Logger LOG = LoggerFactory.getLogger(SolarisNetworkIF.class);
 
     /** Per-interface stats the Solaris HAL reads. Both JNA and FFM concrete subclasses populate this. */
     public static final class IfStats {
@@ -36,12 +28,6 @@ public abstract class SolarisNetworkIF extends AbstractNetworkIF {
         public long inDrops;
         public long speed;
         public long timeStamp;
-    }
-
-    /** Creates a concrete (JNA or FFM) Solaris network interface. */
-    @FunctionalInterface
-    protected interface NetworkIFFactory {
-        SolarisNetworkIF create(NetworkInterface netint) throws InstantiationException;
     }
 
     protected SolarisNetworkIF(NetworkInterface netint) throws InstantiationException {
@@ -67,25 +53,6 @@ public abstract class SolarisNetworkIF extends AbstractNetworkIF {
         this.speed = stats.speed;
         this.timeStamp = stats.timeStamp;
         return true;
-    }
-
-    /**
-     * Gets all network interfaces on this machine, building each with the given concrete factory.
-     *
-     * @param includeLocalInterfaces include local interfaces in the result
-     * @param factory                creates the platform-specific interface instances
-     * @return a list of {@link NetworkIF} objects representing the interfaces
-     */
-    protected static List<NetworkIF> getNetworks(boolean includeLocalInterfaces, NetworkIFFactory factory) {
-        List<NetworkIF> ifList = new ArrayList<>();
-        for (NetworkInterface ni : getNetworkInterfaces(includeLocalInterfaces)) {
-            try {
-                ifList.add(factory.create(ni));
-            } catch (InstantiationException e) {
-                LOG.debug("Network Interface Instantiation failed: {}", e.getMessage());
-            }
-        }
-        return ifList;
     }
 
     /**
